Rename to TypeWrapperStorage. Wrap every Codable type
