--- Source/Shared/Library/PrimitiveWrapper.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-struct PrimitiveWrapper<T: Codable>: Codable {
-  let value: T
-
-  init(value: T) {
-    self.value = value
-  }
-}
