+++ Tests/iOS/Tests/Storage/SyncStorageTests.swift
@@ -8,7 +8,7 @@ final class SyncStorageTests: XCTestCase {
   override func setUp() {
     super.setUp()
     let memory = MemoryStorage(config: MemoryConfig())
-    let primitive = PrimitiveStorage(storage: memory)
+    let primitive = TypeWrapperStorage(storage: memory)
     storage = SyncStorage(storage: primitive)
   }
